+++ packages/datum-service/internal/services/resourcemanager/projects.go
@@ -1,4 +1,4 @@
-package projects
+package resourcemanager
 
 import (
 	"context"
