+++ packages/datum-service/internal/services/resourcemanager/projects_test.go
@@ -1,4 +1,4 @@
-package projects
+package resourcemanager
 
 import (
 	resourcemanagerpb "go.datumapis.com/genproto/os/resourcemanager/v1alpha"
